Separação do conteúdo JS da página gateway_pagamento.php
